requirements.txt
fastapi
azure-search-documents
openai
uvicorn
pytest
